Merge edits to main

+++ fuse_detection/Matrix.cpp
@@ -1,58 +1,57 @@
 #include "Matrix.hpp"
 #include <stdlib.h>
 #include <string.h>
-#include <stdint.h>
 #include "MemoryFree.h"
 // FIXME: Not commenting in this file since I believe Ege said that no dynamic memory will be used. Will
 // comment later
 Matrix::Matrix(const int &num_rows, const int &num_cols){
   failed = false;
   #ifdef ARDUINO
-  if(freeMemory() < (num_rows * sizeof(uint16_t*) + num_rows * num_cols * sizeof(uint16_t))){
+  if(freeMemory() < (num_rows * sizeof(float*) + num_rows * num_cols * sizeof(float))){
     failed = true;
     return;
   }
   #endif
   this->num_rows = num_rows;
   this->num_cols = num_cols;
-  
-  data = malloc(num_rows * sizeof(uint16_t*));
+
+  data = malloc(num_rows * sizeof(float*));
   for(int i = 0; i < num_rows; i++){
-    data[i] = malloc(num_cols * sizeof(uint16_t));
+    data[i] = malloc(num_cols * sizeof(float));
     for(int j = 0; j < num_cols; j++){
       data[i][j] = 0;
     }
   }
 }
 
-Matrix::Matrix(const int &num_rows, const int &num_cols, uint16_t**data){
+Matrix::Matrix(const int &num_rows, const int &num_cols, float **data){
   failed = false;
   #ifdef ARDUINO
-  if(freeMemory() < (num_rows * sizeof(uint16_t*) + num_rows * num_cols * sizeof(uint16_t))){
+  if(freeMemory() < (num_rows * sizeof(float*) + num_rows * num_cols * sizeof(float))){
     failed = true;
     return;
   }
   #endif
   this->num_rows = num_rows;
   this->num_cols = num_cols;
-  
+
   this->data = data;
 }
 
 Matrix::Matrix(const Matrix& m){
   failed = false;
   #ifdef ARDUINO
-  if(freeMemory() < (num_rows * sizeof(uint16_t*) + num_rows * num_cols * sizeof(uint16_t))){
+  if(freeMemory() < (num_rows * sizeof(float*) + num_rows * num_cols * sizeof(float))){
     failed = true;
     return;
   }
   #endif
   num_rows = m.num_rows;
   num_cols = m.num_cols;
-  
-  data = malloc(m.num_rows * sizeof(uint16_t*));
+
+  data = malloc(m.num_rows * sizeof(float*));
   for(int i = 0; i < m.num_rows; i++){
-    data[i] = malloc(m.num_cols * sizeof(uint16_t));
+    data[i] = malloc(m.num_cols * sizeof(float));
     for(int j = 0; j < m.num_cols; j++){
       data[i][j] = m.data[i][j];
     }
@@ -84,7 +83,7 @@ Matrix & Matrix::operator-(const Matrix &rhs){
   return *this;
 }
 
-Matrix & Matrix::operator*(const uint16_t &i){
+Matrix & Matrix::operator*(const float &i){
   for(int row = 0; row < num_rows; row++){
     for(int col = 0; col < num_cols; col++){
       data[row][col] *= i;
@@ -95,7 +94,7 @@ Matrix & Matrix::operator*(const uint16_t &i){
 
 Matrix & Matrix::operator*(const Matrix &rhs){
   for(int row = 0; row < num_rows; row++){
-      uint16_t this_row[num_cols];
+    float this_row[num_cols];
     for(int col = 0; col < num_cols; col++){
       for(int i = 0; i < rhs.num_rows; i++){
         this_row[i] = data[row][col] * rhs.data[row][i];
@@ -134,7 +133,7 @@ bool Matrix::operator!=(const Matrix &rhs){
   return false;
 }
 
-uint16_t & Matrix::cell(int row, int col){
+float & Matrix::cell(int row, int col){
   return data[row][col];
 }
